requirements.txt
django
mysqlclient
djangorestframework
django-cacheops
django-rest-auth
channels
request
